+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.16.0] - 2025-04-15
+## [0.16.0] - 2025-04-22
 
 ### Changes
 
@@ -313,4 +313,4 @@
 [0.14.0]:https://gitlab.ethz.ch/tb/orcai/-/tags/v0.14.0
 [0.15.0]:https://gitlab.ethz.ch/tb/orcai/-/tags/v0.15.0
 [0.15.1]:https://gitlab.ethz.ch/tb/orcai/-/tags/v0.15.1
-[0.16.0]:https://gitlab.ethz.ch/tb/orcai/-/tags/v0.15.2
+[0.16.0]:https://gitlab.ethz.ch/tb/orcai/-/tags/v0.16.0